Gikk bort fra krypterings-util for å tilrettelegge for samme utførelse som tidligere. Fikset noen navn, går fra nokkeler til keys
